notification fragment removed & home, dashboard fragments renamed
