+++ app/src/main/java/com/template/mapapplication/ui/map/MapViewModel.kt
@@ -1,13 +1,13 @@
-package com.template.mapapplication.ui.notifications
+package com.template.mapapplication.ui.map
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class MapViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
