+++ app/src/main/java/com/template/mapapplication/ui/places/PlacesViewModel.kt
@@ -1,10 +1,10 @@
-package com.template.mapapplication.ui.dashboard
+package com.template.mapapplication.ui.places
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class PlacesViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is dashboard Fragment"
